+++ app/src/main/java/io/multy/ui/fragments/send/AssetSendFragment.java
@@ -152,7 +152,7 @@ public class AssetSendFragment extends BaseFragment {
         viewModel.thoseAddress.setValue(inputAddress.getText().toString());
         ((AssetSendActivity) getActivity()).setFragment(R.string.send_from, R.id.container, WalletChooserFragment.newInstance());
         if (getActivity().getIntent().hasCategory(Constants.EXTRA_SENDER_ADDRESS)) {
-            RealmManager.getAssetsDao().getWalletById(getActivity().getIntent().getIntExtra(Constants.EXTRA_WALLET_ID, 0));
+            RealmManager.getAssetsDao().getWalletById(getActivity().getIntent().getLongExtra(Constants.EXTRA_WALLET_ID, 0));
             ((AssetSendActivity) getActivity()).setFragment(R.string.transaction_fee, R.id.container, TransactionFeeFragment.newInstance());
         }
     }
